fix(spark): pass GravitinoPaimonCatalog to ProcedureBuilder in loadProcedure
Delegate loadProcedure to SparkProcedures.newBuilder with
GravitinoPaimonCatalog as the table catalog instead of the internal
Paimon SparkCatalog, so that PaimonProcedureResolver's catalog
validation succeeds when CALL is invoked through Gravitino.
